pyproject.toml
[project]
name = "atb-mcp"
version = "0.1.0"
description = "MCP server for ATB bus data analysis via BigQuery"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0",
    "google-cloud-bigquery>=3.25",
    "pyarrow>=17.0",
    "db-dtypes>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.hatch.build.targets.wheel]
packages = ["src/atb_mcp"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
